restructured App.js for react router, added fullscreen background slideshow
