+++ client/src/pages/LoginRegisterPage.css
@@ -1,0 +1,60 @@
+body {
+  background-color: rgba(28,28,30,0.9);
+  color: rgb(255,255,255);
+  height: 100%;
+}
+
+.temp{
+  position: relative;
+  z-index: 2;
+}
+
+.slideshow{
+  list-style: none;
+  z-index: 1;
+}
+
+.slideshow li span{
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0px;
+  left: 0px;
+  color: transparent;
+  background-size: cover;
+  background-position: 50% 50%;
+  background-repeat: none;
+  opacity: 0;
+  z-index: 0;
+  animation: imageAnimation 24s linear infinite 0s;
+}
+
+.slideshow li:nth-child(1) span{
+  background-image:url(../imgs/mta1.png);
+  filter: blur(8px);
+  -webkit-filter: blur(8px);
+}
+.slideshow li:nth-child(2) span{
+  background-image:url(../imgs/mta2.png);
+  filter: blur(8px);
+  -webkit-filter: blur(8px);
+  animation-delay: 8s;
+}
+.slideshow li:nth-child(3) span{
+  background-image:url(../imgs/mta3.png);
+  filter: blur(8px);
+  -webkit-filter: blur(8px);
+  animation-delay: 16s;
+}
+
+@keyframes imageAnimation {
+  0% { opacity: 0; animation-timing-function: ease-in;}
+  17% { opacity: 1; animation-timing-function: ease-out;}
+  34% { opacity: 1; }
+  51% { opacity: 0; }
+  100% { opacity: 0; }
+}
+
+.no-cssanimations .cb-slideshow li span{
+	opacity: 1;
+}
