+++ client/src/pages/LoginRegisterPage.js
@@ -1,0 +1,32 @@
+import { React, useState } from 'react';
+import './LoginRegisterPage.css';
+
+/*
+https://unsplash.com/photos/PJzeDJAw3oI
+https://unsplash.com/photos/k_j7olQiqAw
+https://unsplash.com/photos/8mswK-LU5Vs
+*/
+
+const LoginRegister = () => {
+  return (
+    <div>
+      <ul className="slideshow">
+        <li>
+          <span>p1</span>
+        </li>
+        <li>
+          <span>p2</span>
+        </li>
+        <li>
+          <span>p3</span>
+        </li>
+      </ul>
+
+      <div className="temp">
+        <h2>Hello</h2>
+      </div>
+    </div>
+  )
+}
+
+export default LoginRegister;
